requirements.txt
streamlit
python-dotenv
PyMuPDF
transformers
sentence-transformers
scikit-learn
langchain
